Website/backend/requirements.txt
transformers
torch
pillow
fastapi
python-multipart
uvicorn
numpy
scipy